Crate p3_goldilocks

The prime field known as Goldilocks, defined as F_p where p = 2^64 - 2^32 + 1.

Structs

Goldilocks

The prime field known as Goldilocks, defined as F_p where p = 2^64 - 2^32 + 1.

MdsMatrixGoldilocks

Poseidon2ExternalLayerGoldilocks

The external layers of the Poseidon2 permutation.

Poseidon2ExternalLayerGoldilocksHL

The external layers of the Poseidon2 permutation used by Horizen Labs.

Poseidon2InternalLayerGoldilocks

The internal layers of the Poseidon2 permutation.

SmallConvolveGoldilocks

Instantiate convolution for “small” RHS vectors over Goldilocks.

Constants

HL_GOLDILOCKS_8_EXTERNAL_ROUND_CONSTANTS

HL_GOLDILOCKS_8_INTERNAL_ROUND_CONSTANTS

MATRIX_DIAG_8_GOLDILOCKS

MATRIX_DIAG_12_GOLDILOCKS

MATRIX_DIAG_16_GOLDILOCKS

MATRIX_DIAG_20_GOLDILOCKS

Type Aliases

Poseidon2Goldilocks

An implementation of the Poseidon2 hash function for the Goldilocks field.

Poseidon2GoldilocksHL

A recreating of the Poseidon2 implementation by Horizen Labs for the Goldilocks field.